pyproject.toml
[tool.uv.workspace]
members = ["common"]

[project]
name = "spring26"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "scipy",
    "torch",
    "torchvision",
    "torchaudio",
    "jupyter",
    "ipykernel",
    "python-dotenv",
]
